+++ AVOS/AVOSCloudIM/AVIMCommon.m
@@ -8,14 +8,23 @@
 
 #import "AVIMCommon.h"
 
-/* AVOSCloud IM code key */
-NSString *const kAVIMCodeKey = @"code";
-/* AVOSCloud IM appCode key */
-NSString *const kAVIMAppCodeKey = @"appCode";
-/* AVOSCloud IM reason key */
-NSString *const kAVIMReasonKey  = @"reason";
-/* AVOSCloud IM detail key */
-NSString *const kAVIMDetailKey  = @"detail";
+// MARK: - Error
+
+NSString * const kAVIMCodeKey = @"code";
+NSString * const kAVIMAppCodeKey = @"appCode";
+NSString * const kAVIMReasonKey  = @"reason";
+NSString * const kAVIMDetailKey  = @"detail";
+
+// MARK: - Conversation
+
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyLastMessage = @"lastMessage";
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyLastMessageAt = @"lastMessageAt";
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyLastReadAt = @"lastReadAt";
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyLastDeliveredAt = @"lastDeliveredAt";
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyUnreadMessagesCount = @"unreadMessagesCount";
+AVIMConversationUpdatedKey AVIMConversationUpdatedKeyUnreadMessagesMentioned = @"unreadMessagesMentioned";
+
+// MARK: - Signature
 
 AVIMSignatureAction AVIMSignatureActionOpen = @"open";
 AVIMSignatureAction AVIMSignatureActionStart = @"start";
@@ -23,3 +32,8 @@ AVIMSignatureAction AVIMSignatureActionAdd = @"invite";
 AVIMSignatureAction AVIMSignatureActionRemove = @"kick";
 AVIMSignatureAction AVIMSignatureActionBlock = @"block";
 AVIMSignatureAction AVIMSignatureActionUnblock = @"unblock";
+
+// MARK: - Deprecated
+
+NSString * const AVIMUserOptionUseUnread = @"AVIMUserOptionUseUnread";
+NSString * const AVIMUserOptionCustomProtocols = @"AVIMUserOptionCustomProtocols";
