--- AVOS/AVOSCloudIM/AVIMOptions.h
@@ -1,19 +1,0 @@
-//
-//  AVIMOptions.h
-//  AVOS
-//
-//  Created by Tang Tianyong on 29/07/2017.
-//  Copyright © 2017 LeanCloud Inc. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface AVIMOptions : NSObject
-
-@property (nonatomic, copy, nullable) NSString *RTMServer;
-
-@end
-
-NS_ASSUME_NONNULL_END
