--- AVOS/AVOSCloudIM/AVIMOptions.m
@@ -1,13 +1,0 @@
-//
-//  AVIMOptions.m
-//  AVOS
-//
-//  Created by Tang Tianyong on 29/07/2017.
-//  Copyright © 2017 LeanCloud Inc. All rights reserved.
-//
-
-#import "AVIMOptions.h"
-
-@implementation AVIMOptions
-
-@end
