--- AVOS/AVOSCloudIM/AVIMUserOptions.m
@@ -1,12 +1,0 @@
-//
-//  AVIMUserOptions.m
-//  AVOS
-//
-//  Created by Tang Tianyong on 8/18/15.
-//  Copyright (c) 2015 LeanCloud Inc. All rights reserved.
-//
-
-#import "AVIMUserOptions.h"
-
-NSString *const AVIMUserOptionUseUnread = @"AVIMUserOptionUseUnread";
-NSString *const AVIMUserOptionCustomProtocols = @"AVIMUserOptionCustomProtocols";
