+++ AVOS/AVOSCloudIM/AVOSCloudIM.m
@@ -8,16 +8,19 @@
 
 #import "AVOSCloudIM.h"
 
+@implementation AVIMOptions
+
+@end
+
 @implementation AVOSCloudIM
 
-+ (AVIMOptions *)defaultOptions {
++ (AVIMOptions *)defaultOptions
+{
     static AVIMOptions *options;
     static dispatch_once_t onceToken;
-
     dispatch_once(&onceToken, ^{
         options = [[AVIMOptions alloc] init];
     });
-
     return options;
 }
